requirements.txt
confluent-kafka==2.5.0
pydantic==2.8.2
pylint==3.2.6
fastapi==0.111.1
uvicorn==0.30.3
dnspython~=2.6.1
idna~=3.7
httpcore~=1.0.5
httpx~=0.27.0
anyio~=4.4.0
h11~=0.14.0
pip~=23.2.1
Jinja2~=3.1.4
dill~=0.3.8
typing_extensions~=4.12.2
rich~=13.7.1
Pygments~=2.18.0
PyYAML~=6.0.2
sniffio~=1.3.1
click~=8.1.7
certifi~=2024.7.4
isort~=5.13.2
colorama~=0.4.6
mdurl~=0.1.2
typer~=0.12.3
shellingham~=1.5.4
MarkupSafe~=2.1.5
astroid~=3.2.4
tomlkit~=0.13.0
mccabe~=0.7.0
platformdirs~=4.2.2
starlette~=0.37.2
email_validator~=2.2.0
pydantic_core~=2.20.1
websockets~=12.0
httptools~=0.6.1
watchfiles~=0.23.0
python-multipart~=0.0.9
python-dotenv~=1.0.1